python-mcp-clients-and-agents/requirements.txt
langchain-mcp-adapters 
langgraph 
langchain-google-vertexai 
langchain-core 
mcp 
pydantic
langchain 
python-dotenv
